lab6
