--- lab6/YourGithubUsername/main.go
@@ -1,21 +1,0 @@
-package main
-
-import (
-	"fmt"
-	"syscall/js"
-)
-
-func CheckPrime(this js.Value, i []js.Value) interface{} {
-	/* add code here */
-}
-
-func registerCallbacks() {
-	js.Global().Set("CheckPrime", js.FuncOf(CheckPrime))
-}
-
-func main() {
-	fmt.Println("Golang main function executed")
-	registerCallbacks()
-
-	//need block the main thread forever
-}
